changed tensor type

+++ result_testing/test_insertion_and_deletion.py
@@ -44,7 +44,7 @@ def insertion(image_tensor,insertion_proportion,device,saliency_map=None):
         insertion_mask = np.where(saliency_map>=cap,1,0)
 
     insertion_mask = np.array([insertion_mask]*3) #insertion mask as (3,w,h)
-    insertion_mask = torch.from_numpy(insertion_mask).to(device) #to torch & to device
+    insertion_mask = torch.from_numpy(insertion_mask).type(torch.FloatTensor).to(device) #to torch & to device
     insertion_mask = insertion_mask.unsqueeze(0)  # Add a batch dimension
 
     return torch.mul(image_tensor,insertion_mask)
@@ -126,11 +126,11 @@ if __name__ == '__main__':
 
                         #random insertion
                         random_insertion = insertion(im2,insertion_proportion,device,saliency_map=None)
-                        score_random.append(np.float64(get_cossine_sim(im1,random_insertion.type(torch.FloatTensor), resnets[i]).cpu().numpy()[0]))
+                        score_random.append(np.float64(get_cossine_sim(im1,random_insertion, resnets[i]).cpu().numpy()[0]))
 
                         #importance map insertion
                         importance_insertion = insertion(im2,insertion_proportion,device,saliency_map=importance_map)
-                        score_importance.append(np.float64(get_cossine_sim(im1,importance_insertion.type(torch.FloatTensor), resnets[i]).cpu().numpy()[0]))
+                        score_importance.append(np.float64(get_cossine_sim(im1,importance_insertion, resnets[i]).cpu().numpy()[0]))
 
                     plt.plot(args.proportion_list,score_random,label = 'random')
                     plt.plot(args.proportion_list,score_importance,label = 'salency_map')
